+++ cmd/launcher/gui/gui_linux.go
@@ -1,10 +1,25 @@
 package gui
 
+// TODO: Implement
+
 func centerWindow(handle uintptr) {
+}
+
+func getWindowDimensions(handle uintptr) (w, h int) {
+	return 200, 100
+}
+
+func setWindowDimensions(handle uintptr, w, h int) {
+}
+
+func flashWindow(handle uintptr) {
 }
 
 func applyIconToWindow(handle uintptr) {
 }
 
 func applyWindowStyle(handle uintptr) {
+}
+
+func setProgressState(s progressState) {
 }