+++ cmd/launcher/gui/gui_windows.h
@@ -1,4 +1,5 @@
 #include <windows.h>
+#include <Commctrl.h>
 
 int centerWindow(ULONG_PTR ulpWindowHandle) {
     HANDLE windowHandle = (HANDLE)ulpWindowHandle;
@@ -18,7 +19,43 @@ int centerWindow(ULONG_PTR ulpWindowHandle) {
     LONG windowHeight = rect.bottom - rect.top;
     LONG newX = (monitorInfo.rcWork.right - windowWidth) / 2;
     LONG newY = (monitorInfo.rcWork.bottom - windowHeight) / 2;
-    SetWindowPos(windowHandle, HWND_TOP, newX, newY, 0, 0, SWP_NOACTIVATE | SWP_NOOWNERZORDER | SWP_NOSIZE);
+    SetWindowPos(windowHandle, HWND_TOP, newX, newY, 0, 0, SWP_NOACTIVATE | SWP_NOZORDER | SWP_NOOWNERZORDER | SWP_NOSIZE);
+}
+
+int getWindowWidth(ULONG_PTR ulpWindowHandle) {
+    HANDLE windowHandle = (HANDLE)ulpWindowHandle;
+    RECT r;
+    BOOL result = GetWindowRect(windowHandle, &r);
+    if (result == 0) {
+        return 200;
+    }
+    return r.right - r.left;
+}
+
+int getWindowHeight(ULONG_PTR ulpWindowHandle) {
+    HANDLE windowHandle = (HANDLE)ulpWindowHandle;
+    RECT r;
+    BOOL result = GetWindowRect(windowHandle, &r);
+    if (result == 0) {
+        return 100;
+    }
+    return r.bottom - r.top;
+}
+
+int setWindowDimensions(ULONG_PTR ulpWindowHandle, int w, int h) {
+    HANDLE windowHandle = (HANDLE)ulpWindowHandle;
+    SetWindowPos(windowHandle, HWND_TOP, 0, 0, (LONG)w, (LONG)h, SWP_NOACTIVATE | SWP_NOZORDER | SWP_NOOWNERZORDER | SWP_NOMOVE);
+}
+
+int flashWindow(ULONG_PTR ulpWindowHandle) {
+    HANDLE windowHandle = (HANDLE)ulpWindowHandle;
+    FLASHWINFO fwi;
+    fwi.cbSize = sizeof(fwi);
+    fwi.hwnd = windowHandle;
+    fwi.dwFlags = FLASHW_TRAY;
+    fwi.uCount = 2;
+    fwi.dwTimeout = 0;
+    FlashWindowEx(&fwi);
 }
 
 HICON largeIcon = NULL;
@@ -48,4 +85,9 @@ void applyIconToWindow(ULONG_PTR ulpWindowHandle) {
 void applyWindowStyle(ULONG_PTR ulpWindowHandle) {
     HANDLE windowHandle = (HANDLE)ulpWindowHandle;
     SetWindowLong(windowHandle, GWL_STYLE, GetWindowLong(windowHandle, GWL_STYLE)&~(WS_SIZEBOX|WS_MAXIMIZEBOX));
+}
+
+void setProgressBarState(ULONG_PTR ulpBarHandle, int progressState) {
+    HWND barHWND = (HWND)ulpBarHandle;
+    SendMessage(barHWND, 1040, (WPARAM)progressState, 0);
 }