+++ pkg/launcher/config/deployment_config.go
@@ -18,8 +18,9 @@ type DeploymentConfig struct {
 }
 
 type HashDataConfig struct {
-	BundleInfoURL string `json:"BundleInfoURL"`
-	BaseURL       string `json:"BaseURL,omitempty"`
+	BundleInfoURL     string `json:"BundleInfoURL"`
+	BaseURL           string `json:"BaseURL,omitempty"`
+	IsUpdateMandatory bool   `json:"IsUpdateMandatory,omitempty"`
 }
 
 type LauncherUpdateConfig struct {
@@ -54,7 +55,7 @@ func (dc *DeploymentConfig) GetLauncherUpdateConfig() *LauncherUpdateConfig {
 	if dc.HasLauncherUpdateConfig() {
 		return &dc.LauncherUpdate[0]
 	}
-	panic(fmt.Sprintf("Deployment config had %d launcher updates. Expected 1.", len(dc.LauncherUpdate)))
+	return nil
 }
 
 func ReadDeploymentConfig(reader io.Reader, os string, arch string) (string, error) {